obj handling

+++ test_system/apis/organizations/views.py
@@ -1,3 +1,4 @@
+from django.shortcuts import get_object_or_404
 from test_system.apps.organizations.models import Organization
 from test_system.apis.organizations.serializers import OrganizationSerializer
 from django.http import Http404
@@ -29,26 +30,22 @@ class OrganizationCreateView(APIView):
 class OrganizationGetUpdateDeleteView(APIView):
     permission_classes= [OrganizationPermission]
 
-    def get_object(self, organization_id):
-        try:
-            return Organization.objects.get(id=organization_id)
-        except Organization.DoesNotExist:
-            return Http404
-
     def get(self, request, organization_id, format=None):
-        organization = self.get_object(organization_id)
+        organization = get_object_or_404(Organization, id=organization_id)
+        self.check_object_permissions(request, organization)
         serializer = OrganizationSerializer(organization)
         return Response(serializer.data)
 
     def put(self, request, organization_id, format=None):
-        organization = self.get_object(organization_id)
+        organization = get_object_or_404(Organization, id=organization_id)
+        self.check_object_permissions(request, organization)
         serializer = OrganizationSerializer(organization, data=request.data)
-        if serializer.is_valid():
+        if serializer.is_valid(raise_exception=True):
             serializer.save()
             return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, organization_id):
-        organization = self.get_object(organization_id)
+        organization = get_object_or_404(Organization, id=organization_id)
+        self.check_object_permissions(request, organization)
         organization.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
